# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "电商订单支付系统 API 文档",
        "title": "Order Payment System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/home/address": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "地址"
                ],
                "summary": "创建地址",
                "parameters": [
                    {
                        "description": "地址信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_internal_types.AddressReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            }
        },
        "/home/address/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "地址"
                ],
                "summary": "地址列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/order-payment-system_internal_types.AddressResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/home/address/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "地址"
                ],
                "summary": "更新地址",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "地址ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "地址信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_internal_types.AddressReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "地址"
                ],
                "summary": "删除地址",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "地址ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            }
        },
        "/home/cart/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "添加到购物车",
                "parameters": [
                    {
                        "description": "商品信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_internal_types.AddCartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            }
        },
        "/home/cart/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "删除购物车商品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "商品ID",
                        "name": "goods_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            }
        },
        "/home/cart/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "购物车列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/order-payment-system_internal_types.CartListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/home/cart/select": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "切换购物车勾选",
                "parameters": [
                    {
                        "description": "勾选信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_internal_types.ToggleSelectReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            }
        },
        "/home/cart/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "更新购物车数量",
                "parameters": [
                    {
                        "description": "更新信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_internal_types.UpdateCartReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "order-payment-system_internal_types.AddCartReq": {
            "type": "object",
            "required": [
                "goods_id",
                "quantity"
            ],
            "properties": {
                "goods_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "order-payment-system_internal_types.AddressReq": {
            "type": "object",
            "required": [
                "detail",
                "phone",
                "receiver"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "detail": {
                    "type": "string",
                    "maxLength": 200
                },
                "district": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 20
                },
                "province": {
                    "type": "string"
                },
                "receiver": {
                    "type": "string",
                    "maxLength": 50
                }
            }
        },
        "order-payment-system_internal_types.AddressResp": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_default": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "receiver": {
                    "type": "string"
                }
            }
        },
...
        "order-payment-system_internal_types.CartItemResp": {
            "type": "object",
            "properties": {
                "goods_id": {
                    "type": "integer"
                },
                "goods_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "selected": {
                    "type": "boolean"
                }
            }
        },
        "order-payment-system_internal_types.CartListResp": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/order-payment-system_internal_types.CartItemResp"
                    }
                },
                "total_price": {
                    "description": "仅计算selected的",
                    "type": "integer"
                }
            }
        },
...
        "order-payment-system_internal_types.ToggleSelectReq": {
            "type": "object",
            "required": [
                "goods_id"
            ],
            "properties": {
                "goods_id": {
                    "type": "integer"
                },
                "selected": {
                    "type": "boolean"
                }
            }
        },
        "order-payment-system_internal_types.UpdateCartReq": {
            "type": "object",
            "required": [
                "goods_id",
                "quantity"
            ],
            "properties": {
                "goods_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "order-payment-system_pkg_response.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}